CRUD nas telas dos produtos OK

+++ Edecasa/Models/Produto.cs
@@ -15,8 +15,8 @@ namespace Edecasa.Models
         public int Id { get; set; }
         public string Descricao { get; set; }
         public string Categoria { get; set; }
-        public float VlPequeno { get; set; }
-        public float VlGrande { get; set; }
+        public double VlPequeno { get; set; }
+        public double VlGrande { get; set; }
         public ICollection<Item> Itens { get; set; }
     }
 }
